Matrix functions: moved to integration tests
